refactor: keep the route harness inside the widget that needs it
`RouteHarness` sat in the shared `widgets/` folder with no production caller and
one widget's tests and stories as its only consumers, and it duplicated the
router boilerplate in `widgets/stories/with-story-router.tsx`. Move it next to
that widget until a second one needs it, and note there that growing the shared
story router a `/` route beats keeping a second copy.

`withStoryRouteSearch` was a six-line wrapper used by one stories file, which
also reached for the harness directly. It folds into that file.
